Add custom hook for adding the filter for the dates and checkbox
